requirements.txt
fastapi[all]
sqlalchemy
alembic
pydantic
python-dotenv
uvicorn
psycopg2
pytest
httpx